pull components into smaller files & start type defs
