+++ app/components/views/Offset/ProjectSearch/Checkbox/styles.ts
